+++ src/utils/directoryScroll.ts
@@ -191,8 +191,9 @@ const DirectoryScrollService = {
    * @returns A promise that resolves when the scroll operation is complete
    */
   scrollToSetting: async (): Promise<void> => {
-    // Find and scroll to the campaign element using the active class
-    await DirectoryScrollService.scrollToElement('.fcb-setting-folder.folder:not(.collapsed)');
+    // Find and scroll to the setting element using the active class
+    // This is simpler than the others because it's the top folder (no never hidden) and there's always only one
+    await DirectoryScrollService.scrollToElement('.fcb-setting-folder.folder .folder-header');
   },
 
   /**